Issue #1191 Avoid running ohmyzsh install more than once and correct the git URL (#1192)
* Issue #1191 Avoid running ohmyzsh install more than once and correct the git URL

* Issue #1191 Touch a file separate from the cloned oh my zsh source that flags oh my zsh as installed

* Issue #1191 Change owner of the flag file for oh my zsh installation

+++ scripts/install-ohmyzsh.sh
@@ -2,7 +2,16 @@
 
 # Install oh-my-zsh
 
-git clone git://github.com/robbyrussell/oh-my-zsh.git /home/vagrant/.oh-my-zsh
+if [ -f /home/vagrant/.ohmyzsh ]
+then
+    echo "oh my zsh already installed."
+    exit 0
+fi
+
+touch /home/vagrant/.ohmyzsh
+chown -Rf vagrant:vagrant /home/vagrant/.ohmyzsh
+
+git clone https://github.com/robbyrussell/oh-my-zsh.git /home/vagrant/.oh-my-zsh
 cp /home/vagrant/.oh-my-zsh/templates/zshrc.zsh-template /home/vagrant/.zshrc
 printf "\nemulate sh -c 'source ~/.bash_aliases'\n" | tee -a /home/vagrant/.zprofile
 printf "\nemulate sh -c 'source ~/.profile'\n" | tee -a /home/vagrant/.zprofile
